+++ orbit/utils/common_test.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Tests for orbit.utils."""
+"""Tests for orbit.utils.common."""
 
-from orbit import utils
+from orbit.utils import common
 
 import tensorflow as tf
 
@@ -22,12 +22,13 @@ import tensorflow as tf
 class UtilsTest(tf.test.TestCase):
 
   def test_create_global_step(self):
-    step = utils.create_global_step()
+    step = common.create_global_step()
+    self.assertEqual(step.name, "global_step:0")
     self.assertEqual(step.dtype, tf.int64)
     self.assertEqual(step, 0)
     step.assign_add(1)
     self.assertEqual(step, 1)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
   tf.test.main()
